// lib/types.ts
export interface Funnel {
  id: string;
  name: string;
  lineIconUrl?: string;
  adVideoImageUrl?: string;
  adVideoUrls?: string[];
  adCtas?: string[];
  adLpImageUrl?: string;
  adLpLinks?: string[];
  tpImageUrl?: string;
  tpLinks?: string[];
  lineRegisterImageUrl?: string;
  lineDeliveryImageUrl?: string;
  distributionUrls?: string[];
  createdAt: string;
  updatedAt: string;
}